# docs/openapi.yml
openapi: 3.0.3
info:
  title: Attendee API
  version: 1.0.0
  description: Meetings bots made easy
paths:
...
  /api/v1/calendar_events:
    get:
      operationId: List Calendar Events
      description: Returns a list of calendar events for the authenticated project.
        Results are paginated using cursor pagination.
      summary: List calendar events
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      - in: query
        name: calendar_deduplication_key
        schema:
          type: string
        description: Filter events by calendar deduplication key
        examples:
          DeduplicationKeyExample:
            value: user-abcd
            summary: Deduplication Key Example
      - in: query
        name: calendar_id
        schema:
          type: string
        description: Filter events by calendar ID
        examples:
          CalendarIDExample:
            value: cal_abcdef1234567890
            summary: Calendar ID Example
      - in: query
        name: cursor
        schema:
          type: string
        description: Cursor for pagination
      - in: query
        name: updated_after
        schema:
          type: string
        description: Filter events updated after this timestamp (ISO 8601 format)
        examples:
          UpdatedAfterExample:
            value: '2025-01-13T10:30:00Z'
            summary: Updated After Example
      tags:
      - Calendars
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCalendarEventList'
              examples:
                CalendarEvent:
                  value:
                    next: http://api.example.org/accounts/?cursor=cD00ODY%3D"
                    previous: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
                    results:
                    - id: evt_abcdef1234567890
                      calendar_id: cal_abcdef1234567890
                      platform_uuid: google_event_123456789
                      meeting_url: https://meet.google.com/abc-defg-hij
                      name: Event Name
                      start_time: '2025-01-15T14:00:00Z'
                      end_time: '2025-01-15T15:00:00Z'
                      is_deleted: false
                      attendees:
                      - email: user1@example.com
                        name: John Doe
                      - email: user2@example.com
                        name: Jane Smith
                      raw:
                        google_event_data: '...'
                      bots:
                      - id: bot_abcdef1234567890
                        metadata:
                          customer_id: abc123
                        meeting_url: https://meet.google.com/abc-defg-hij
                        state: joined_recording
                        events: []
                        transcription_state: complete
                        recording_state: complete
                        join_at: '2025-01-15T14:00:00Z'
                      created_at: '2025-01-13T10:30:00.123456Z'
                      updated_at: '2025-01-13T10:30:00.123456Z'
                  summary: Calendar Event
                  description: Example of a calendar event with associated bots
          description: List of calendar events
  /api/v1/calendars:
    get:
      operationId: List Calendars
      description: Returns a list of calendars for the authenticated project. Results
        are paginated using cursor pagination.
      summary: List calendars
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      - in: query
        name: cursor
        schema:
          type: string
        description: Cursor for pagination
      - in: query
        name: deduplication_key
        schema:
          type: string
        description: Filter calendars by deduplication key
        examples:
          DeduplicationKeyExample:
            value: user-abcd
            summary: Deduplication Key Example
      tags:
      - Calendars
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCalendarList'
          description: List of calendars
    post:
      operationId: Create Calendar
      description: After being created, the calendar will be connected to the specified
        calendar platform.
      summary: Create a new calendar
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      tags:
      - Calendars
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarRequest'
            examples:
              CreateGoogleCalendar:
                value:
                  client_id: 123456789-abcdefghijklmnopqrstuvwxyz.apps.googleusercontent.com
                  client_secret: GOCSPX-abcdefghijklmnopqrstuvwxyz
                  refresh_token: 1//04abcdefghijklmnopqrstuvwxyz
                  platform: google
                  metadata:
                    tenant_id: '1234567890'
                  deduplication_key: user-abcd
                summary: Create Google Calendar
                description: Example of creating a Google calendar connection
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
              examples:
                NewlyCreatedCalendar:
                  value:
                    id: cal_abcdef1234567890
                    platform: google
                    state: connected
                    metadata:
                      tenant_id: '1234567890'
                    deduplication_key: user-abcd
                    connection_failure_data: null
                    created_at: '2025-01-13T10:30:00.123456Z'
                    updated_at: '2025-01-13T10:30:00.123456Z'
                  summary: Newly Created Calendar
                  description: Example response when a calendar is successfully created
          description: Calendar created successfully
        '400':
          description: Invalid input
  /api/v1/calendars/{object_id}:
    get:
      operationId: Get Calendar
      description: Returns the details of a specific calendar.
      summary: Get calendar details
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      - in: path
        name: object_id
        schema:
          type: string
        description: Calendar ID
        required: true
        examples:
          CalendarIDExample:
            value: cal_abcdef1234567890
            summary: Calendar ID Example
      tags:
      - Calendars
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
              examples:
                NewlyCreatedCalendar:
                  value:
                    id: cal_abcdef1234567890
                    platform: google
                    state: connected
                    metadata:
                      tenant_id: '1234567890'
                    deduplication_key: user-abcd
                    connection_failure_data: null
                    created_at: '2025-01-13T10:30:00.123456Z'
                    updated_at: '2025-01-13T10:30:00.123456Z'
                  summary: Newly Created Calendar
                  description: Example response when a calendar is successfully created
          description: Calendar details
        '404':
          description: Calendar not found
    patch:
      operationId: Update Calendar
      description: Updates calendar credentials (client_secret, refresh_token) or
        metadata.
      summary: Update calendar
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      - in: path
        name: object_id
        schema:
          type: string
        description: Calendar ID
        required: true
        examples:
          CalendarIDExample:
            value: cal_abcdef1234567890
            summary: Calendar ID Example
      tags:
      - Calendars
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPatchCalendarRequest'
            examples:
              UpdateCredentials:
                value:
                  client_secret: GOCSPX-NewClientSecret123
                  refresh_token: 1//05o3zfluegTFVCgYICGHGAUSNgF-L9Ir23dcclPCJW7KmzPhsQaNFcAzNwQkV6uM1gIGID8nBelYDPtbIr123
                summary: Update credentials
                description: Example of updating calendar credentials
              UpdateMetadataOnly:
                value:
                  metadata:
                    department: sales
                    team: frontend
                    updated: 'true'
                summary: Update metadata only
                description: Example of updating only the calendar metadata
              UpdateRefreshTokenOnly:
                value:
                  refresh_token: 1//05NewRefreshTokenHere
                summary: Update refresh token only
                description: Example of updating only the refresh token
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
              examples:
                NewlyCreatedCalendar:
                  value:
                    id: cal_abcdef1234567890
                    platform: google
                    state: connected
                    metadata:
                      tenant_id: '1234567890'
                    deduplication_key: user-abcd
                    connection_failure_data: null
                    created_at: '2025-01-13T10:30:00.123456Z'
                    updated_at: '2025-01-13T10:30:00.123456Z'
                  summary: Newly Created Calendar
                  description: Example response when a calendar is successfully created
          description: Calendar updated successfully
        '400':
          description: Invalid input
        '404':
          description: Calendar not found
    delete:
      operationId: Delete Calendar
      description: Permanently deletes a calendar and all associated data.
      summary: Delete calendar
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
          default: Token YOUR_API_KEY_HERE
        description: API key for authentication
        required: true
      - in: header
        name: Content-Type
        schema:
          type: string
          default: application/json
        description: Should always be application/json
        required: true
      - in: path
        name: object_id
        schema:
          type: string
        description: Calendar ID
        required: true
        examples:
          CalendarIDExample:
            value: cal_abcdef1234567890
            summary: Calendar ID Example
      tags:
      - Calendars
      security:
      - {}
      responses:
        '200':
          description: Calendar deleted successfully
        '404':
          description: Calendar not found
components:
  schemas:
    Bot:
      type: object
      properties:
        id:
          type: string
        metadata:
          type: object
          description: Metadata associated with the bot
          readOnly: true
        meeting_url:
          type: string
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/BotStateEnum'
          readOnly: true
        events:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              sub_type:
                type: string
                nullable: true
              created_at:
                type: string
                format: date-time
          readOnly: true
        transcription_state:
          allOf:
          - $ref: '#/components/schemas/TranscriptionStateEnum'
          readOnly: true
        recording_state:
          allOf:
          - $ref: '#/components/schemas/RecordingStateEnum'
          readOnly: true
        join_at:
          type: string
          format: date-time
        deduplication_key:
          type: string
      required:
      - deduplication_key
      - events
      - id
      - join_at
      - meeting_url
      - metadata
      - recording_state
      - state
      - transcription_state
...
    BotStateEnum:
      type: string
      enum:
      - ready
      - joining
      - joined_not_recording
      - joined_recording
      - leaving
      - post_processing
      - fatal_error
      - waiting_room
      - ended
      - data_deleted
      - scheduled
      - staged
      - joined_recording_paused
      - joining_breakout_room
      - leaving_breakout_room
    Calendar:
      type: object
      properties:
        id:
          type: string
        platform:
          allOf:
          - $ref: '#/components/schemas/PlatformEnum'
          readOnly: true
        client_id:
          type: string
          readOnly: true
        platform_uuid:
          type: string
          readOnly: true
          nullable: true
        state:
          allOf:
          - $ref: '#/components/schemas/CalendarStateEnum'
          readOnly: true
        metadata:
          type: object
          description: Metadata associated with the calendar
          readOnly: true
        deduplication_key:
          type: string
          readOnly: true
          nullable: true
          description: Optional key for deduplicating calendars
        connection_failure_data:
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        last_successful_sync_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        last_attempted_sync_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - client_id
      - connection_failure_data
      - created_at
      - deduplication_key
      - id
      - last_attempted_sync_at
      - last_successful_sync_at
      - metadata
      - platform
      - platform_uuid
      - state
      - updated_at
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        calendar_id:
          type: string
        platform_uuid:
          type: string
          readOnly: true
        meeting_url:
          type: string
          readOnly: true
          nullable: true
        start_time:
          type: string
          format: date-time
          readOnly: true
        end_time:
          type: string
          format: date-time
          readOnly: true
        is_deleted:
          type: boolean
          readOnly: true
        attendees:
          readOnly: true
          nullable: true
        ical_uid:
          type: string
          readOnly: true
          nullable: true
        name:
          type: string
          readOnly: true
          nullable: true
        raw:
          readOnly: true
        bots:
          type: array
          items:
            $ref: '#/components/schemas/Bot'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - attendees
      - bots
      - calendar_id
      - created_at
      - end_time
      - ical_uid
      - id
      - is_deleted
      - meeting_url
      - name
      - platform_uuid
      - raw
      - start_time
      - updated_at
    CalendarStateEnum:
      type: string
      enum:
      - connected
      - disconnected
...
    CreateCalendarRequest:
      type: object
      properties:
        platform_uuid:
          type: string
          minLength: 1
          description: The UUID of the calendar on the calendar platform. Specify
            only for non-primary calendars.
        client_id:
          type: string
          minLength: 1
          description: The client ID for the calendar platform authentication
        client_secret:
          type: string
          minLength: 1
          description: The client secret for the calendar platform authentication
        refresh_token:
          type: string
          minLength: 1
          description: The refresh token for accessing the calendar platform
        platform:
          allOf:
          - $ref: '#/components/schemas/PlatformEnum'
          description: |-
            The calendar platform (google or microsoft)

            * `google` - Google
            * `microsoft` - Microsoft
        metadata:
          description: JSON object containing metadata to associate with the calendar
        deduplication_key:
          type: string
          minLength: 1
          description: Optional key for deduplicating calendars. If a calendar with
            this key already exists in the project, the new calendar will not be created
            and an error will be returned.
      required:
      - client_id
      - client_secret
      - platform
      - refresh_token
...
    PaginatedCalendarEventList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEvent'
    PaginatedCalendarList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Calendar'
...
    PatchedPatchCalendarRequest:
      type: object
      properties:
        client_secret:
          type: string
          minLength: 1
          description: The client secret for the calendar platform authentication
        refresh_token:
          type: string
          minLength: 1
          description: The refresh token for accessing the calendar platform
        metadata:
          description: JSON object containing metadata to associate with the calendar
    PlatformEnum:
      enum:
      - google
      - microsoft
      type: string
      description: |-
        * `google` - Google
        * `microsoft` - Microsoft
...
    RecordingStateEnum:
      type: string
      enum:
      - not_started
      - in_progress
      - complete
      - failed
      - paused
...
    TranscriptionStateEnum:
      type: string
      enum:
      - not_started
      - in_progress
      - complete
      - failed
...
servers:
- url: https://app.attendee.dev
  description: Production server
...